manager/requirements.txt
dj_database_url==0.5.0
dj-stripe==2.3.0
django-allauth==0.41.0
django-configurations==2.2
django-extensions==2.2.9
django-intercom==0.1.3
django-prometheus==2.0.0
django-rest-knox==4.1.0
django-sendgrid-v5==0.8.1
Django==3.0.6
djangorestframework-camel-case==1.1.2
djangorestframework==3.11.0
drf-yasg==1.17.1
google-api-python-client==1.8.3
pyjwt==1.7.1
psycopg2-binary==2.8.5
sentry-sdk==0.14.3
